Upload the mod

+++ Oxygen/sonic3air/source/sonic3air/client/archipelago/ArchipelagoClient.cpp
@@ -63,7 +63,7 @@ void ArchipelagoClient::updateConnection(float timeElapsed)
 	mConnecting = (state > APClient::State::DISCONNECTED && state < APClient::State::SLOT_CONNECTED);
 }
 
-void ArchipelagoClient::sendLocation(uint32 id)
+void ArchipelagoClient::sendLocation(uint64 id)
 {
 	std::list<int64_t> idList;
 	idList.push_front(id);
